search_api/requirements.txt
django==1.9.8
django-environ==0.4.0
djangorestframework==3.4.0
elasticsearch==2.3.0
elasticsearch_dsl
django-cors-headers

